Run Welford accumulators in float32
The streaming Welford accumulators (total_sum, rate_mean, rate_M2) in
_compute_stats_from_stream were allocated as float64 out of caution.
Empirically that's overkill for KPF stack sizes:

  11-frame bias stack on real 20240923 data, float32 vs float64 Welford:
    GREEN_IMG  max_abs_diff = 3.8e-6 ADU  (vs ~5 e- read noise)
    RED_IMG    max_abs_diff = 3.8e-6 ADU
    GREEN_SNR  max_abs_diff = 9.5e-7
    RED_SNR    max_abs_diff = 1.9e-6
    median abs diff: 0

The differences are at the float32 representation limit — seven orders
of magnitude below per-pixel read noise. float32 Welford only starts to
diverge at N >> 1e4 frames, far beyond any realistic KPF calibration
stack.

Saves ~800 MB of transient memory during masters generation
(12 arrays × 4080² × 4 bytes saved).

+++ kpfpipe/modules/masters/base.py
@@ -402,9 +402,9 @@ class BaseMasterModule:
 
                 exact_stats[ext] = {}
                 exact_stats[ext]['nframe'] = np.zeros((NROW,NCOL),dtype=np.int32)
-                exact_stats[ext]['total_sum'] = np.zeros((NROW,NCOL),dtype=np.float64)
-                exact_stats[ext]['rate_mean'] = np.zeros((NROW,NCOL),dtype=np.float64)
-                exact_stats[ext]['rate_M2'] = np.zeros((NROW,NCOL),dtype=np.float64)
+                exact_stats[ext]['total_sum'] = np.zeros((NROW,NCOL),dtype=np.float32)
+                exact_stats[ext]['rate_mean'] = np.zeros((NROW,NCOL),dtype=np.float32)
+                exact_stats[ext]['rate_M2'] = np.zeros((NROW,NCOL),dtype=np.float32)
 
                 approx_mean = approx_stats[ext]['rate_mean']
                 approx_rms = approx_stats[ext]['rate_rms']
